services/telegram/pyproject.toml
[project]
name = "telegram-service"
version = "0.1.0"
description = "Telegram microservice for mee6"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-telegram-bot>=20.0",
    "pydantic",
    "httpx",
    "psycopg2-binary",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
